Imported game fonts and setup the UI system to show player life, timer and game over text.

+++ Assets/Scripts/Enemies/EnemyController.cs
@@ -19,7 +19,7 @@ public class EnemyController : BaseGameObject
 
     public float shotProbability;
 
-    public float shotDamage;
+    public int shotDamage;
 
     public float life;
